+++ zombienet/examples/custom-script.js
@@ -1,0 +1,7 @@
+const assert = require("assert");
+
+async function run(nodeName, networkInfo, args) {
+    return 2084;
+}
+
+module.exports = { run }